Separates templates from classes
