+++ src/components/film-top-rated.js
@@ -1,12 +1,6 @@
 import {createElement} from '../utils.js';
+import {createFilmTopRatedContainerTemplate} from '../templates/film-top-raited.js';
 
-const createFilmTopRatedContainerTemplate = () => {
-  return (`<section class="films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-      <div class="films-list__container">
-      </div>
-    </section>`);
-};
 
 export default class FilmTopRated {
   constructor() {
