+++ src/components/films-container.js
@@ -1,12 +1,6 @@
 import {createElement} from '../utils.js';
+import {createFilmsContainerTemplate} from '../templates/films-container.js';
 
-const createFilmsContainerTemplate = () => {
-  return (`<section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container">
-        </div>
-      </section>`);
-};
 
 export default class FilmsContainer {
   constructor() {
