+++ src/components/statistics.js
@@ -1,13 +1,5 @@
 import {createElement} from './../utils.js';
-
-const createFooterStatTemplate = (moviesCount) => {
-  let numberFormat = new Intl.NumberFormat(`ru`);
-  numberFormat = numberFormat.format(moviesCount);
-
-  return (
-    `<p>${numberFormat} movies inside</p>`
-  );
-};
+import {createFooterStatTemplate} from '../templates/statistics.js';
 
 export default class Statistics {
   constructor(amount) {
