+++ src/templates/button-show-more.js
@@ -1,0 +1,7 @@
+const createButtonShowMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createButtonShowMoreTemplate};
