+++ src/templates/film-card.js
@@ -1,0 +1,27 @@
+const createFilmCardTemplate = (film) => {
+  const {title, poster, rating, productionDate, duration, genres, comments, description} = film;
+
+  const longDescriptionFormat = description.length < 140 ? description : description.slice(0, 139) + `...`;
+
+  return (
+    `<article class="film-card">
+          <h3 class="film-card__title">${title}</h3>
+          <p class="film-card__rating">${rating}</p>
+          <p class="film-card__info">
+            <span class="film-card__year">${productionDate}</span>
+            <span class="film-card__duration">${duration}</span>
+            <span class="film-card__genre">${genres[0]}</span>
+          </p>
+          <img src="${poster}" alt="" class="film-card__poster">
+          <p class="film-card__description">${longDescriptionFormat}</p>
+          <a class="film-card__comments">${comments.length} comments</a>
+          <form class="film-card__controls">
+            <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
+            <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
+            <button class="film-card__controls-item button film-card__controls-item--favorite">Mark as favorite</button>
+          </form>
+        </article>`
+  );
+};
+
+export {createFilmCardTemplate};
