+++ src/templates/films.js
@@ -1,0 +1,5 @@
+const createFilmsTemplate = () => {
+  return `<section class="films"></section>`;
+};
+
+export {createFilmsTemplate};
