+++ src/templates/no-films.js
@@ -1,0 +1,9 @@
+const createNoFilmsTemplate = () => {
+  return (
+    `<section class="films-list">
+      <h2 class="films-list__title">There are no movies in our database</h2>
+    </section>`
+  );
+};
+
+export {createNoFilmsTemplate};
